Add files via upload
Task 1: Assign the number (1 to 4) below the image of the rectangle with respect to its
length inside the rectangle. The shorter the line lower the number (No need to reorder the
image of the rectangle, only give numbering)

Task 2: Align(make the rectangle image straight) all the given images of the rectangle.